Adds login on signup, and logout (MH-58)
